refactor(exif): unify value converters
